# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fixit 24/7 API - On-demand home repair marketplace
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type RegisterBodyRole =
  (typeof RegisterBodyRole)[keyof typeof RegisterBodyRole];

export const RegisterBodyRole = {
  homeowner: "homeowner",
  tradie: "tradie",
} as const;

export interface RegisterBody {
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterBodyRole;
  phone?: string;
  suburb?: string;
  postcode?: string;
  bio?: string;
  skills?: number[];
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  homeowner: "homeowner",
  tradie: "tradie",
  admin: "admin",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  phone?: string | null;
  suburb?: string | null;
  postcode?: string | null;
  bio?: string | null;
  avatarUrl?: string | null;
  rating?: number | null;
  reviewCount: number;
  isActive: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UpdateProfileBody {
  name?: string;
  phone?: string;
  suburb?: string;
  postcode?: string;
  bio?: string;
  avatarUrl?: string;
  skills?: number[];
}

export type AdminUpdateUserBodyRole =
  (typeof AdminUpdateUserBodyRole)[keyof typeof AdminUpdateUserBodyRole];

export const AdminUpdateUserBodyRole = {
  homeowner: "homeowner",
  tradie: "tradie",
  admin: "admin",
} as const;

export interface AdminUpdateUserBody {
  name?: string;
  role?: AdminUpdateUserBodyRole;
  isActive?: boolean;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  icon: string;
  description?: string | null;
  jobCount: number;
}

export interface CreateCategoryBody {
  name: string;
  icon: string;
  description?: string;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  open: "open",
  matched: "matched",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type JobUrgency = (typeof JobUrgency)[keyof typeof JobUrgency];

export const JobUrgency = {
  standard: "standard",
  urgent: "urgent",
  emergency: "emergency",
} as const;

export interface Job {
  id: number;
  title: string;
  description: string;
  status: JobStatus;
  urgency: JobUrgency;
  categoryId: number;
  categoryName?: string | null;
  homeownerId: number;
  homeownerName?: string | null;
  suburb?: string | null;
  postcode?: string | null;
  address?: string | null;
  imageUrls: string[];
  budget?: number | null;
  claimCount: number;
  createdAt: string;
  scheduledFor?: string | null;
}

export type ClaimStatus = (typeof ClaimStatus)[keyof typeof ClaimStatus];

export const ClaimStatus = {
  pending: "pending",
  accepted: "accepted",
  rejected: "rejected",
  withdrawn: "withdrawn",
  completed: "completed",
} as const;

export interface Claim {
  id: number;
  jobId: number;
  tradieId: number;
  tradieName?: string | null;
  tradieRating?: number | null;
  tradieReviewCount: number | null;
  tradieSuburb?: string | null;
  tradieAvatarUrl?: string | null;
  status: ClaimStatus;
  message?: string | null;
  proposedPrice?: number | null;
  createdAt: string;
}

export type JobDetail = Job & {
  claims?: Claim[];
  matchedTradies?: User[];
};

export interface JobListResponse {
  jobs: Job[];
  total: number;
  page: number;
  limit: number;
}

export type CreateJobBodyUrgency =
  (typeof CreateJobBodyUrgency)[keyof typeof CreateJobBodyUrgency];

export const CreateJobBodyUrgency = {
  standard: "standard",
  urgent: "urgent",
  emergency: "emergency",
} as const;

export interface CreateJobBody {
  title: string;
  description: string;
  categoryId: number;
  urgency: CreateJobBodyUrgency;
  suburb?: string;
  postcode?: string;
  address?: string;
  imageUrls?: string[];
  budget?: number;
  scheduledFor?: string;
}

export type UpdateJobBodyStatus =
  (typeof UpdateJobBodyStatus)[keyof typeof UpdateJobBodyStatus];

export const UpdateJobBodyStatus = {
  open: "open",
  matched: "matched",
  in_progress: "in_progress",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export type UpdateJobBodyUrgency =
  (typeof UpdateJobBodyUrgency)[keyof typeof UpdateJobBodyUrgency];

export const UpdateJobBodyUrgency = {
  standard: "standard",
  urgent: "urgent",
  emergency: "emergency",
} as const;

export interface UpdateJobBody {
  title?: string;
  description?: string;
  status?: UpdateJobBodyStatus;
  urgency?: UpdateJobBodyUrgency;
  address?: string;
  budget?: number;
  scheduledFor?: string;
}

export interface CreateClaimBody {
  message?: string;
  proposedPrice?: number;
}

export type UpdateClaimBodyStatus =
  (typeof UpdateClaimBodyStatus)[keyof typeof UpdateClaimBodyStatus];

export const UpdateClaimBodyStatus = {
  accepted: "accepted",
  rejected: "rejected",
  withdrawn: "withdrawn",
  completed: "completed",
} as const;

export interface UpdateClaimBody {
  status: UpdateClaimBodyStatus;
}

export interface Notification {
  id: number;
  userId: number;
  type: string;
  title: string;
  message: string;
  isRead: boolean;
  jobId?: number | null;
  createdAt: string;
}

export interface UnreadCountResponse {
  count: number;
}

export interface HomeownerDashboard {
  totalJobs: number;
  openJobs: number;
  inProgressJobs: number;
  completedJobs: number;
  totalSpent: number;
  recentJobs: Job[];
  pendingClaims: number;
}

export interface TradieDashboard {
  activeJobs: number;
  completedJobs: number;
  totalEarnings: number;
  availableLeads: number;
  myRating?: number | null;
  recentClaims: Claim[];
  availableJobs: Job[];
}

export interface AdminDashboard {
  totalUsers: number;
  totalHomeowners: number;
  totalTradies: number;
  totalJobs: number;
  openJobs: number;
  completedJobs: number;
  totalCategories: number;
  recentJobs: Job[];
  recentUsers: User[];
}

export type ListJobsParams = {
  status?: string;
  categoryId?: number;
  urgency?: string;
  page?: number;
  limit?: number;
};

export type ListNotificationsParams = {
  unreadOnly?: boolean;
};

export type AdminListUsersParams = {
  role?: string;
  page?: number;
  limit?: number;
};

export type AdminListJobsParams = {
  status?: string;
  page?: number;
  limit?: number;
};